+++ src/service/caller.jsx
@@ -16,4 +16,15 @@ Axios.interceptors.request.use(request => {
     return request
 })
 
+Axios.interceptors.response.use(response => {
+    return response;
+}, error => {
+    if(error.response.status === 401){
+        accountService.logout()
+        window.location = "/auth/login"
+    } else {
+        return Promise.reject(error)
+    }
+})
+
 export default Axios;
